+++ module-utils/time/time_conversion.hpp
@@ -137,6 +137,7 @@ namespace utils
 
             UTF8 get_date_time_substr(GetParameters param);
             uint32_t get_date_time_sub_value(GetParameters param);
+            uint32_t get_UTC_date_time_sub_value(GetParameters param);
         };
 
         /// helper class to operate on time now